+++ src/main/java/life/majiang/community/model/User.java
@@ -2,14 +2,15 @@ package life.majiang.community.model;
 
 import lombok.Data;
 
+import java.util.List;
+
 @Data
 public class User {
-    private Integer id;
+    private Long id;
     private String name;
     private String accountId;
     private String token;
     private Long gmtCreate;
     private Long gmtModified;
     private String avatarUrl;
-
 }
